app/vmselect/promql: fix panic with using @ with series which is not present at the start of the query
Previously, "selector @ another_selector" assumed that "another_selector" metric is supposed to exist since "start" used in the query.

If the query was evaluated in the following case:
- start - 2, end - 10
- timestamps for "another_selector" 5,6,7,8,9,10
- timestamps for "selector"
The resulting "at" timestamp would be taken from NaN, causing a panic later.

This commit changes this and checks for the first non-NaN value. This makes it easier to use for users as series are not always aligned and returning an error in this case would disallow using this for some time ranges.

Signed-off-by: Zakhar Bessarab <z.bessarab@victoriametrics.com>