docs: Fix the definition of active time series
Based on the code, an active time series is one that has received at least one
sample during the last hour. It does not includes querying.

The first time the docs were updated with this was in d06aae9454bf808545c5e3b7d7e98f71a0f8839a.
But even at that time, the code shows that querying a metric did not make it active.

Signed-off-by: Artem Fetishev <rtm@victoriametrics.com>