hourMetricIDCache: also record the indexDB id in a cache to speed up data ingestion
There are two instances of hourMetricIDCache hmPrev and hmCurr. Previously, each
data ingestion call (Storage.add()) needed to retrieve the indexDB for both in
order to determine if they belong to the same indexDB. This slowed down the
insertion.

In order to eliminate this idb id retrieval, we populate it at the time of the
cache creation.

Signed-off-by: Artem Fetishev <rtm@victoriametrics.com>