docs/cloud: migrate docs to cloud repo
This commit deletes the cloud docs folder after moving it into the
private cloud repo. The main reason is to keep things tidy and
handle reviews in a better way, as the helm charts repo is doing.

Future updates should be protected since the github actions file
with rsync command should not be messing with this folder when
updating everything.