lib/storage: Change BenchmarkHeadPostingForMatchers to use global index time range explicitly
During the data retrieval, VictoriaMetrics switches to global index search if
the time range is > 40 days. Prior ba0d7dc2fc7fd3d0d1f9fc51121d9044f7175172,
this switch was handled in indexDB code. That commit moved that logic to the
storage level. As the result, indexDB will use search per-day index for whatever
time range that is passed to it.

This broke the BenchmarkHeadPostingForMatchers. It now shows very slow indexDB
performance compared to v1.111.0 (the last release before that commit). This is
because now indexDB tries to search 55 years of data by spawning a separate
goroutine for each day. Even though the data was inserted only for the current
day, running that many goroutines significantly slows down the search.

The fix is to use global index time range explicitly.

Signed-off-by: Artem Fetishev <rtm@victoriametrics.com>