app/vmctl: print import error before pipe error
See https://github.com/VictoriaMetrics/VictoriaMetrics/issues/9220#issuecomment-3001880833:
```
2025-06-24T17:27:47.062Z      error VictoriaMetrics/app/vmctl/backoff/backoff.go:60 got error: failed to write into "http://vminsert-vmcluster.metrics.svc.cluster.local:8480/insert/0/prometheus": io: read/write on closed pipe on attempt: 1; will retry in 2s
```

This error could happen only if pipe-reader was closed before we tried to write into pipe-writer.
But in this case we won't see the error explaining why pipe-reader was closed. In this change,
we checking if there was a pipe-reader error before checking pipe-writer error.

Signed-off-by: hagen1778 <roman@victoriametrics.com>