issue-7599: do not force read-only for legacy idbs (#9355)
`isReadOnly` flag does not prevent writes (which we need to perform in
case of deletion), it just [disables background
merges](https://github.com/VictoriaMetrics/VictoriaMetrics/issues/2603).
In some cases, it could lead to a large number of small parts and hurt
performance.

Union of legacy deleted metric ids is already propagated to partitioned
indexDBs. We also need to add them to the previous legacy indexDB.