Rewrite tests so they are not flaky
Previously tests opened idbs but did not close them. This led to their bg tasks
running even after the test completion. Once the test is completed its working
directory is removed while the bg tasks still try to access them and panic.

The fix is to create idbs by opening the storage and closing the storage after
the test completion.

Signed-off-by: Artem Fetishev <rtm@victoriametrics.com>