app/vmctl: allow overriding tmp files location for prometheus migration
Previously, vmctl tried to create a tmp directory in a directory with Prometheus snapshot. This is not always possible as snapshot can be mounted in read-only mode.

Use a system default temporary location and allow user to customize the tmp path in order to avoid issues with custom tmp locations.

See: https://github.com/VictoriaMetrics/VictoriaMetrics/issues/9505
Signed-off-by: Zakhar Bessarab <z.bessarab@victoriametrics.com>