lib/storage: fix double counting in vm_deleted_metrics_total
The vm_deleted_metrics_total metric value represents the number of metricIDs
stored in deletedMetricIDs cache. This cache lives at the storage level and
stores the deleted metrics from both prev and curr idbs. However, the metric
is populated at the idb level. Since there are always 2 idbs (prev and curr),
the value is populated twice. Hence the doubled value of the metric.

The fix is to populate the metric value at the storage level.

Signed-off-by: Artem Fetishev <rtm@victoriametrics.com>