lib/storage: introduce metricNameSearch type
Searching metricName by metricID happens many times during a single API call.
This requires getting the current set of idbs before those calls happen.
Which is fine but requires propagating idbs across the code base. This is also
fine in case of OSS version as it is used in Search only.

Propagating idbs across the code base becomes a problem in Enterprise version
as it is used in at least 3 places. As a result it becomes very difficult to
merge things from OSS to Ent.

Localizing the all the dependencies in one searchMetricName type and reusing
this type everywhere should make things simpler.

Signed-off-by: Artem Fetishev <rtm@victoriametrics.com>