deployment/docker: bump base image version
Follow up on
https://github.com/VictoriaMetrics/VictoriaMetrics/pull/9805

the image tag didn’t change in #9805. So Docker could reuse a cached
image. This required manual cleanup with docker image rm or docker
system prune to force building the fixed image.
With this change, the image tag is updated, guaranteeing that the new
image with the security fix is used everywhere.