app/vmselect: add more context for bad URL error
It was multiple times already for users to get confused with Single-node
and Cluster URL formats for VictoriaMetrics. This is an attempt to bring
more context to the error message, if request doesn't contain expected
"/select" or "/delete" prefix. Which is usually a sign of using wrong
API.

This commit is only an attempt to fix it. But it demonstrates the idea.

Related PR https://github.com/VictoriaMetrics/VictoriaMetrics/pull/9741