lib/envflag: apply -secret.flags inside envflag.Parse function
Follow up on PR:
https://github.com/VictoriaMetrics/VictoriaMetrics/pull/9839

and addresses review comment
https://github.com/VictoriaMetrics/VictoriaMetrics/pull/9839#discussion_r2477729886

Alex: this design decision isn't good, since it will lead to potential
security issues over time when we'll forget adding ApplySecretFlags()
call after the flag.Parse() call or add it at the wrong place. BTW, we
do not call flag.Parse() explicitly - instead envflag.Parse() is called.
So it is natural to call ApplySecretFlags() inside this call. Are there
restrictions which prevent from doing this? If there are no
restrictions, then there is no need in making this function public - it
will be called explicitly inside envflag.Parse().