lib/storage: Fix data race in containsTimeRange()
When one goroutine attemps to update the min timestamp under the lock it could
have been updated already by another goroutine with a smaller timestamp. As a
result the goroutine will update the timestamp with a bigger value.

A simple unit test (included in this commit) demonstrates that.

Additionally, use a simple Mutex instead of RWMutex.
RWMutexes only introduce an unnecessary overhead for operations as simple as
retrieving a value from a map and regular Mutex should be preferred.

Thanks to @valyala for spotting a bug and the advice on RWMutexes.

Signed-off-by: Artem Fetishev <rtm@victoriametrics.com>