lib/protoparser/opentelemetry: use separate loggers for unsupported delta temporality/metric type logs.
A throttled logger will continue to log messages occasionally with a suffix indicating how many similar logs were throttled.  Using the same logger for multiple log messages can result in certain logs being entirely suppressed and invisible in the logs.
This updates most of the loggers used in `appendFromScopeMetrics` to be their own logger so that "unsupported delta temporality/metric type" logs will be visible for all metric types.
Additionally, `skippedSampleLogger` is only used by `appendSamplesFromHistogram` so this was moved closer to that function.

Related to #9447
Related to #9498