lib/storage: add indexDB block cache metric unit test
indexDB has 3 block caches. These caches export metrics. Storage collects these
metrics for each indexDB it has (currently prev and curr only).

There is a potential problem:
- These caches are shared by all indexDBs
- Each indexDB reports the block cache metrics.
- Storage collects the metrics of all indexDBs by adding them together.

I.e. it is possible to count block cache metrics several times.
It is not the case in current implementation because the addition of the metrics
is not performed intentionally.

The added unit test 1) demonstrates that the resulting counts are reported
correctly and 2) protects from future unintentional changes in this behavior.

Signed-off-by: Artem Fetishev <rtm@victoriametrics.com>