lib/storage: also create parts.json on parition creation
Currently, when a partition is created its corresponding parts.json file is not
created right away (see createNewParition()). Its creation is delayed until the
first part files are created on disk (see swapSrcWithDstParts()). However, the
parts.json file is created for a possibly empty partition when an existing
partition is opened (see mustOpenPartition()) and when a partition snapshot is
create (see MustCreateSnapshotAt()).

I.e. `parts.json` is an important part of a partition, since it is an artifact
that describes the partition contents. And it should be created on pt creation
even if its contents is empty.

Signed-off-by: Artem Fetishev <rtm@victoriametrics.com>