influx/parser{,_test}.go: account for excess white spaces in write str
nimon adds excess white spaces in the write string and does not set a
timestamp. This confuses VictoriaMetrics because it treats the empty
string as the timestamp.

Add a test case for this situation, and ensure that the conversion to
an integer is only done if the string actually contains something.

See-also: https://sourceforge.net/p/nmon/bugs/32/
Fixes: https://github.com/VictoriaMetrics/VictoriaMetrics/issues/10049
Signed-off-by: Nowa Ammerlaan <nowa.ammerlaan@surf.nl>