lib/storage: share one instance of dateMetricIDCache among all indexDBs
This is how it always has been until 73a1ce2dd6 which moved this cache to
indexDB. I.e. each indexDB had its own instance of cache.

This makes sense from the design point of view, such stronger encapsulation
(each indexDB maintains its own cache for its own data).

However, the cache has the memory size limit. And since we now have 2 cache
instances (for curr and prev idbs), this size is doubled. With swithing to
partition index, the will be N time bigger (where N is the number of
partitions). It does not automatically mean that the total memory consumption
will be N time bigger, the cache consumes no memory if it is unused. However
if someone performs a backfill for all N partitions, it will be possible that
consumption will be N time bigger. This may be fine for the duration of the
backfill operation, but then the caches for all partitions will remain unused
but full until next restart. So some additional cleanup logic would need to
be implemented.

But even when used in typical condictions (when the database receives only very
recent data) there are problems with reporting the cache stats. Since we now
have more than one cache its stats is aggregated. I.e. the sizeBytes and
sizeMaxBytes are summed. And the utilization can only be calculated for all the
instances: totalSizeBytes/totalSizeMaxBytes. But this way it is impossible to
see if a given cache instance is full and take necessary action.

Of course we could now report the cache stats for every indexDB, without
aggregation. But this will double the number of metrics now and will result in
N-times more metrics with partition index. Even if this okay too, the standard
monitoring dashboards still show the aggregated stats.

So having one instance of dateMetricID cache will:

1) allow to have predictable cache size (does not depend on number of indexDBs)
2) simpler cache logic (no eviction based on how old its entries)
3) make the monitoring simpler.

Signed-off-by: Artem Fetishev <rtm@victoriametrics.com>