lib/storage: fix dateMetricIDCache EntriesCount() and SizeBytes() stats
Previously the cache reported this stats for its immutable part only. Whereas
there are cases when the number of entries in its mutable part is comparable
with the number in immutable part.

The stats from the mutable part remains invisible until it is sync'ed to the
immutable part. It is also possible that the cache gets reset after the sync
because the cache size exceeds the max allowed size.

Reporting the stats for both mutable and immutable parts should provide a clear
picture of the cache utilization and take approviate actions if necessary (such
as adding more memory or adjusting the cache limits).

Signed-off-by: Artem Fetishev <rtm@victoriametrics.com>